TINKERPOP-2727 Allow for a null value to has()
Seems to apply when the null is on the first arg of a two arg overload - it tries to call on has(T,...) which ends in IllegalArgumentException, which never quite surfaces to the JavaTranslator properly. Another example of why bytecode is lacking and why gremlin-language is likely going to be its successor.